requirements.txt
Django>=4.2.0
Pillow>=10.0.0
opencv-python-headless>=4.8.0
face-recognition>=1.3.0
numpy>=1.24.0
python-decouple>=3.8
django-crispy-forms>=2.0
crispy-bootstrap5>=0.7
